2D Twin Stick Shooter game
A 2D Twin-stick Shooter with multiple levels switched based on the scenes. The first level is a person in planet Uranus shooting at enemies, with a Mini Boss Spawning after a fixed time. The second level is a space battle with ships and a captain ship, where the objective is to collect the lost ship's wreckages.